Implement copy method and apply.
